requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
werkzeug==2.3.0
pyotp==2.8.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
webauthn>=2.0.0,<3.0.0
